+++ Second/First/Second.asmx.cs
@@ -88,7 +88,7 @@ namespace Second
         }
 
         [WebMethod]
-        public long CadastrarUsuario(int alID, String asUserId,String asUUID,byte[] asFoto,String asNome)
+        public long CadastrarUsuario(long alID, String asUserId,String asUUID,byte[] asFoto,String asNome)
         {
             return iControleUsuarios.cadastrarUsuario(alID, asUserId, asUUID, asFoto, asNome);
         }
